Revert "feat: phoatis_intent_cls pass test_seacrowd"
This reverts commit 71ee4f18ccd21be72e48f2356ab33962df5c5da8.
