--- seacrowd/sea_datasets/phoatis_intent_cls/intent_label.txt
@@ -1,29 +1,0 @@
-UNK
-abbreviation
-aircraft
-aircraft#flight
-aircraft#flight#flight_no
-airfare
-airfare#flight
-airfare#flight_time
-airline
-airline#flight
-airline#flight_no
-airport
-capacity
-city
-city#flight_time
-distance
-flight
-flight#flight_no
-flight#flight_time
-flight_no
-flight_no#flight_time
-flight_time
-ground_fare
-ground_fare#ground_service
-ground_service
-meal
-quantity
-restriction
-day_name
