--- seacrowd/sea_datasets/phoatis_intent_cls/slot_label.txt
@@ -1,142 +1,0 @@
-PAD
-UNK
-O
-B-aircraft_code
-B-airline_code
-B-airline_name
-I-airline_name
-B-airport_code
-B-airport_name
-I-airport_name
-B-arrive_date.date_relative
-I-arrive_date.date_relative
-B-arrive_date.day_name
-I-arrive_date.day_name
-B-arrive_date.day_number
-I-arrive_date.day_number
-B-arrive_date.month_name
-I-arrive_date.month_name
-B-arrive_date.today_relative
-B-arrive_time.end_time
-I-arrive_time.end_time
-B-arrive_time.period_mod
-I-arrive_time.period_mod
-B-arrive_time.period_of_day
-I-arrive_time.period_of_day
-B-arrive_time.start_time
-I-arrive_time.start_time
-B-arrive_time.time
-I-arrive_time.time
-B-arrive_time.time_relative
-I-arrive_time.time_relative
-B-city_name
-I-city_name
-B-class_type
-I-class_type
-B-connect
-I-connect
-B-cost_relative
-I-cost_relative
-B-day_name
-I-day_name
-B-day_number
-I-day_number
-B-days_code
-B-depart_date.date_relative
-I-depart_date.date_relative
-B-depart_date.day_name
-I-depart_date.day_name
-B-depart_date.day_number
-I-depart_date.day_number
-B-depart_date.month_name
-I-depart_date.month_name
-B-depart_date.today_relative
-I-depart_date.today_relative
-B-depart_date.year
-I-depart_date.year
-B-depart_time.end_time
-I-depart_time.end_time
-B-depart_time.period_mod
-I-depart_time.period_mod
-B-depart_time.period_of_day
-I-depart_time.period_of_day
-B-depart_time.start_time
-I-depart_time.start_time
-B-depart_time.time
-I-depart_time.time
-B-depart_time.time_relative
-I-depart_time.time_relative
-B-economy
-I-economy
-B-fare_amount
-I-fare_amount
-B-fare_basis_code
-B-flight_days
-I-flight_days
-B-flight_mod
-I-flight_mod
-B-flight_number
-B-flight_stop
-I-flight_stop
-B-flight_time
-I-flight_time
-B-fromloc.airport_code
-B-fromloc.airport_name
-I-fromloc.airport_name
-B-fromloc.city_name
-I-fromloc.city_name
-B-fromloc.state_code
-B-fromloc.state_name
-I-fromloc.state_name
-B-meal
-I-meal
-B-meal_code
-I-meal_code
-B-meal_description
-I-meal_description
-B-mod
-I-mod
-B-month_name
-B-or
-B-period_of_day
-I-period_of_day
-B-restriction_code
-I-restriction_code
-B-return_date.date_relative
-I-return_date.date_relative
-B-return_date.day_name
-I-return_date.day_name
-B-return_date.day_number
-I-return_date.day_number
-B-return_date.month_name
-I-return_date.month_name
-B-return_date.today_relative
-I-return_date.today_relative
-B-return_time.period_mod
-B-return_time.period_of_day
-I-return_time.period_of_day
-B-round_trip
-I-round_trip
-B-state_code
-B-state_name
-B-stoploc.airport_name
-B-stoploc.city_name
-I-stoploc.city_name
-B-stoploc.state_code
-B-time
-I-time
-B-time_relative
-B-today_relative
-I-today_relative
-B-toloc.airport_code
-B-toloc.airport_name
-I-toloc.airport_name
-B-toloc.city_name
-I-toloc.city_name
-B-toloc.country_name
-I-toloc.country_name
-B-toloc.state_code
-B-toloc.state_name
-I-toloc.state_name
-B-transport_type
-I-transport_type
